Moved public doc files
